+++ presentation/src/main/java/com/daejol/presentation/category/match/PersonalActivity.kt
@@ -1,4 +1,4 @@
-package com.daejol.presentation.personalMatchScreen
+package com.daejol.presentation.category.match
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
